added u16 opcode support

+++ crates/wasmi/src/engine/bytecode/mod.rs
@@ -6,33 +6,22 @@ mod utils;
 mod tests;
 
 pub use self::utils::{
-    BranchOffset,
-    BranchParams,
-    DataSegmentIdx,
-    DropKeep,
-    DropKeepError,
-    ElementSegmentIdx,
-    FuncIdx,
-    GlobalIdx,
-    LocalDepth,
-    Offset,
-    SignatureIdx,
-    TableIdx,
+    BranchOffset, BranchParams, DataSegmentIdx, DropKeep, DropKeepError, ElementSegmentIdx,
+    FuncIdx, GlobalIdx, LocalDepth, Offset, SignatureIdx, TableIdx,
 };
 use core::fmt::Debug;
 use wasmi_core::UntypedValue;
 
 #[derive(Default, Debug, Copy, Clone, PartialEq, Eq)]
-pub struct InstrMeta(pub usize, pub u8);
+pub struct InstrMeta(pub usize, pub u16);
 
 impl InstrMeta {
-
     pub fn source_pc(&self) -> u32 {
         self.0 as u32
     }
 
-    pub fn opcode(&self) -> u8 {
-        self.1 as u8
+    pub fn opcode(&self) -> u16 {
+        self.1
     }
 }
